+++ packages/shared/package.json
@@ -38,7 +38,6 @@
     "vitest": "catalog:"
   },
   "dependencies": {
-    "@typescript-eslint/utils": "catalog:",
     "minimatch": "^10.0.1",
     "postgres": "catalog:"
   }
